Refactor Modal component to support actor and scene modals, enhancing reusability. Update Actors and Studio pages to utilize the new Modal structure for actor creation and scene management, improving overall UI consistency and functionality.
