+++ src/components/molecules/Modal/BaseModal.tsx
@@ -1,0 +1,76 @@
+import { Fragment } from 'react';
+
+import { Dialog, DialogPanel, DialogTitle, Transition, TransitionChild } from '@headlessui/react';
+import { X } from 'lucide-react';
+
+import { cn } from '../../../utils/cn';
+
+interface BaseModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  title: string;
+  children: React.ReactNode;
+  maxWidth?: 'sm' | 'md' | 'lg' | 'xl';
+}
+
+const maxWidthClasses = {
+  sm: 'max-w-sm',
+  md: 'max-w-md',
+  lg: 'max-w-lg',
+  xl: 'max-w-xl',
+};
+
+export function BaseModal({ isOpen, onClose, title, children, maxWidth = 'md' }: BaseModalProps) {
+  return (
+    <Transition appear show={isOpen} as={Fragment}>
+      <Dialog as='div' className='relative z-50' onClose={onClose}>
+        <TransitionChild
+          as={Fragment}
+          enter='ease-out duration-300'
+          enterFrom='opacity-0'
+          enterTo='opacity-100'
+          leave='ease-in duration-200'
+          leaveFrom='opacity-100'
+          leaveTo='opacity-0'
+        >
+          <div className='fixed inset-0 bg-black/25' />
+        </TransitionChild>
+
+        <div className='fixed inset-0 overflow-y-auto'>
+          <div className='flex min-h-full items-center justify-center p-4 text-center'>
+            <TransitionChild
+              as={Fragment}
+              enter='ease-out duration-300'
+              enterFrom='opacity-0 scale-95'
+              enterTo='opacity-100 scale-100'
+              leave='ease-in duration-200'
+              leaveFrom='opacity-100 scale-100'
+              leaveTo='opacity-0 scale-95'
+            >
+              <DialogPanel
+                className={cn(
+                  'w-full transform overflow-auto rounded-2xl bg-background p-6 text-left align-middle shadow-xl transition-all',
+                  maxWidthClasses[maxWidth],
+                )}
+              >
+                <div className='flex items-center justify-between mb-4'>
+                  <DialogTitle as='h3' className='text-lg font-medium leading-6 text-primary'>
+                    {title}
+                  </DialogTitle>
+                  <button
+                    onClick={onClose}
+                    className='rounded-full p-1 hover:bg-primary/10 transition-colors'
+                  >
+                    <X className='h-5 w-5 text-primary' />
+                  </button>
+                </div>
+
+                {children}
+              </DialogPanel>
+            </TransitionChild>
+          </div>
+        </div>
+      </Dialog>
+    </Transition>
+  );
+}
